Moves meshFunctionTransfer to lf::refinement
